+++ src/setup-docker/templates/fabric-docker/scripts/base-functions.sh
@@ -18,7 +18,7 @@ function certsGenerate() {
 
   if [ -d "$FULL_CERT_PATH" ]; then
     echo "Can't generate certs, directory already exists : $FULL_CERT_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -54,7 +54,7 @@ function genesisBlockCreate() {
 
   if [ -f "$OUTPUT_PATH/$GENESIS_FILE_NAME" ]; then
     echo "Cant't generate genesis block, file already exists: $OUTPUT_PATH/$GENESIS_FILE_NAME"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -87,7 +87,7 @@ function createChannelTx() {
 
   if [ -f "$CHANNEL_TX_PATH" ]; then
     echo "Can't create channel configuration, it already exists : $CHANNEL_TX_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi
 
@@ -122,7 +122,7 @@ function createNewChannelUpdateTx() {
 
   if [ -f "$ANCHOR_PEER_UPDATE_PATH" ]; then
     echo "Cant't create anchor peer update, it already exists : $ANCHOR_PEER_UPDATE_PATH"
-    echo "Try using 'reboot' or 'down' to remove whole network or 'start' to reuse it"
+    echo "Try using 'reset' or 'down' to remove whole network or 'start' to reuse it"
     exit 1
   fi